+++ examples/azure_key/AzureRsaKey.cs
@@ -1,6 +1,7 @@
 using Azure.Security.KeyVault.Keys.Cryptography;
 using System;
 using System.Security.Cryptography;
+using System.Threading;
 using AzureSignatureAlgorithm = Azure.Security.KeyVault.Keys.Cryptography.SignatureAlgorithm;
 
 namespace Tmds.Ssh.AzureKeyExample;
@@ -9,12 +10,17 @@ sealed class AzureRsaKey : RSA
 {
     private readonly CryptographyClient _cryptoClient;
     private readonly RSAParameters _publicParameters;
+    private readonly CancellationToken _cancellationToken;
 
-    public AzureRsaKey(CryptographyClient client, RSAParameters publicParameters)
+    public AzureRsaKey(
+        CryptographyClient client,
+        RSAParameters publicParameters,
+        CancellationToken cancellationToken)
     {
         KeySizeValue = publicParameters.Modulus!.Length * 8;
         _cryptoClient = client;
         _publicParameters = publicParameters;
+        _cancellationToken = cancellationToken;
     }
 
     public override RSAParameters ExportParameters(bool includePrivateParameters)
@@ -44,7 +50,10 @@ sealed class AzureRsaKey : RSA
             _ => throw new CryptographicException($"Unsupported hash algorithm {hashAlgorithm.Name}"),
         };
 
-        byte[] res = _cryptoClient.Sign(sigAlgo, hash).Signature;
-        return res;
+        SignResult res = _cryptoClient.SignAsync(
+            sigAlgo,
+            hash,
+            _cancellationToken).GetAwaiter().GetResult();
+        return res.Signature;
     }
 }
