+++ examples/azure_key/Program.cs
@@ -1,4 +1,10 @@
+using Azure.Core;
+using Azure.Identity;
+using Azure.Security.KeyVault.Keys;
 using System;
+using System.Buffers.Binary;
+using System.Security.Cryptography;
+using System.Text;
 using System.Threading.Tasks;
 
 namespace Tmds.Ssh.AzureKeyExample;
@@ -30,30 +36,52 @@ class Program
         string vaultName = args[1];
         string keyName = args[2];
 
+        DefaultAzureCredential cred = new(includeInteractiveCredentials: true);
+        string keyVaultUrl = $"https://{vaultName}.vault.azure.net/";
+        KeyClient keyClient = new KeyClient(new Uri(keyVaultUrl), cred);
+        KeyVaultKey key = await keyClient.GetKeyAsync(keyName);
+
         if (action == "print_pub_key")
         {
-            return await PrintPublicKey(vaultName, keyName);
+            return PrintPublicKeyAsync(key);
         }
         else
         {
             string destination = args.Length >= 4 ? args[3] : "localhost";
             string command = args.Length >= 5 ? args[4] : "echo 'hello world'";
-            return await SshExec(vaultName, keyName, destination, command);
+            return await SshExecAsync(cred, key, destination, command);
         }
     }
 
-    private static async Task<int> PrintPublicKey(string vaultName, string keyName)
+    private static int PrintPublicKeyAsync(KeyVaultKey key)
     {
-        string pubKey = await AzureKeyCredential.GetAzurePubKey(vaultName, keyName);
+        string pubKey;
+        if (key.KeyType == KeyType.Rsa)
+        {
+            RSAParameters pubParams = key.Key.ToRSA(includePrivateParameters: false)
+                .ExportParameters(false);
+            pubKey = GetRsaPubKey(pubParams);
+        }
+        else if (key.KeyType == KeyType.Ec)
+        {
+            ECParameters pubParams = key.Key.ToECDsa(includePrivateParameters: false)
+                .ExportParameters(false);
+            pubKey = GetEcdsaPubKey(pubParams);
+        }
+        else
+        {
+            throw new NotImplementedException($"Unsupported Azure key type {key.KeyType}");
+        }
+
         Console.WriteLine(pubKey);
         return 0;
     }
 
-    private static async Task<int> SshExec(string vaultName, string keyName, string destination, string command)
+    private static async Task<int> SshExecAsync(TokenCredential credential, KeyVaultKey key, string destination, string command)
     {
         SshClientSettings clientSettings = new SshClientSettings(destination)
         {
-            Credentials = [new AzureKeyCredential(vaultName, keyName)],
+            Credentials = [new AzureKeyCredential(credential, key)],
         };
         using SshClient client = new SshClient(clientSettings);
 
@@ -104,5 +132,57 @@ class Program
                 }
             }
         }
+    }
+
+    private static string GetRsaPubKey(RSAParameters pubParams)
+    {
+        byte[] n = pubParams.Modulus!;
+        byte[] e = pubParams.Exponent!;
+
+        // If the modulus has the highest bit set, we need to pad it with a 0
+        // byte.
+        int padding = 0;
+        if ((n[0] & 0x80) != 0)
+        {
+            padding = 1;
+        }
+
+        Span<byte> keyData = stackalloc byte[4 + 7 + 4 + e.Length + 4 + padding + n.Length];
+        BinaryPrimitives.WriteInt32BigEndian(keyData, 7);
+        Encoding.ASCII.GetBytes("ssh-rsa", keyData.Slice(4));
+        BinaryPrimitives.WriteInt32BigEndian(keyData.Slice(11), e.Length);
+        e.CopyTo(keyData.Slice(15, e.Length));
+        BinaryPrimitives.WriteInt32BigEndian(keyData.Slice(15 + e.Length), n.Length + padding);
+        keyData[19 + e.Length] = 0;
+        n.CopyTo(keyData.Slice(19 + e.Length + padding));
+
+        return $"ssh-rsa {Convert.ToBase64String(keyData)}";
+    }
+
+    private static string GetEcdsaPubKey(ECParameters pubParams)
+    {
+        byte[] x = pubParams.Q.X!;
+        byte[] y = pubParams.Q.Y!;
+
+        string curveName = pubParams.Curve.Oid?.FriendlyName switch
+        {
+            "ECDSA_P256" => "nistp256",
+            "ECDSA_P384" => "nistp384",
+            "ECDSA_P521" => "nistp521",
+            _ => throw new NotImplementedException($"Unsupported ECDSA curve {pubParams.Curve.Oid?.FriendlyName}"),
+        };
+        string keyType = $"ecdsa-sha2-{curveName}";
+
+        Span<byte> keyData = stackalloc byte[4 + keyType.Length + 4 + curveName.Length + 4 + 1 + x.Length + y.Length];
+        BinaryPrimitives.WriteInt32BigEndian(keyData, keyType.Length);
+        Encoding.ASCII.GetBytes(keyType, keyData.Slice(4));
+        BinaryPrimitives.WriteInt32BigEndian(keyData.Slice(4 + keyType.Length), curveName.Length);
+        Encoding.ASCII.GetBytes(curveName, keyData.Slice(8 + keyType.Length));
+        BinaryPrimitives.WriteInt32BigEndian(keyData.Slice(8 + keyType.Length + curveName.Length), x.Length + y.Length + 1);
+        keyData[12 + keyType.Length + curveName.Length] = 0x04;
+        x.CopyTo(keyData.Slice(13 + keyType.Length + curveName.Length));
+        y.CopyTo(keyData.Slice(13 + keyType.Length + curveName.Length + x.Length));
+
+        return $"{keyType} {Convert.ToBase64String(keyData)}";
     }
 }
